Add delete post API in post router

+++ src/routes/post.route.js
@@ -66,11 +66,16 @@ router.get("/:id", isLoggedIn, async(req, res) => {
 router.delete("/:postId", isLoggedIn, async (req, res) => {
     try {
         const { postId } = req.params
+        const foundUser = req.user
 
         const post = await Post.findById(postId)
 
         if (!post) {
            throw new Error("Post not found")
+        }
+
+        if (post.authorId.toString() !== foundUser._id.toString()) {
+            throw new Error("You are not authorized to perform this action")
         }
 
         await Post.findByIdAndDelete(postId)
